Moved spinner out to an extension, so it can easily be used in other examples
